requirements.txt
asgiref==3.4.1
certifi==2021.5.30
Django==3.2.5
djangorestframework==3.12.4
gunicorn==20.1.0
pytz==2021.1
sqlparse==0.4.1
